move useFocusToggle hook
